% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_mouse.R
\name{lazy_mouse}
\alias{lazy_mouse}
\title{Randomly move the mouse cursor}
\usage{
lazy_mouse(
  number_of_repetitions = 30,
  time_between_repetitions = 60,
  pixels = 50
)
}
\arguments{
\item{number_of_repetitions}{Number of repetitions}

\item{time_between_repetitions}{Seconds between repetitions}

\item{pixels}{Mouse movement (macOS only)}
}
\description{
Randomly move the mouse cursor
}
